requirements.txt
ase==3.20.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
community==1.0.0b1
cycler==0.10.0
decorator==4.4.0
Flask==1.0.3
future==0.18.2
googledrivedownloader==0.4
h5py==2.9.0
idna==2.8
imageio==2.5.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.1.0
llvmlite==0.35.0
MarkupSafe==1.1.1
matplotlib==3.1.0
networkx==2.3
numba==0.52.0
numpy==1.16.4
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
plyfile==0.7
protobuf==3.8.0
pydicom==1.2.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-louvain==0.14
pytz==2019.1
PyWavelets==1.0.3
pyzmq==19.0.0
rdflib==4.2.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.1.0
six==1.12.0
tensorboardX==1.7
threadpoolctl==2.1.0
torchfile==0.1.0
-e git+https://github.com/ncullen93/torchsample.git@21507feb258a25bf6924e4844e578624cda72140#egg=torchsample
tornado==6.0.4
tqdm==4.51.0
urllib3==1.25.3
visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==0.15.4
